Generate trace for benchmarks
Command line option --genTrace for generating trace for benchmarks. A new config file called "gen_trace.py" for generating trace is
created under the folder ./configs/DRAMSim2/. An example script to run the trace generation called "genTraceExample" is created in ./ folder
Need to change the TraceBasedSim file to adapt to this special kind of trace.