Fix the termination error at cycle 0. Add gen_trace script to submit jobs for generting traces.
The bug is due to how the tracefile string is handled in DRAMSim2. The tracefile string originally is
used to create a folder, while we use it again to create a file with the same name in the same folder.
This is solved by adding ".trc" for the name of the file so that it does not conflict with the folder name.