Run trace based simulation for SPEC benchmark traces
I modified the TraceBasedSim so that now it can simulate the traces of SPEC
benchmarks which has the relative time format. Now the trace based simulation will
take two traces as input traces, one for each CPU. option -t is for the first trace
and -T is for the second trace.